Create WixTestTools project and reorganize files.
