+++ src/TestBA/Hresult.cs
