+++ src/WixTestTools/BundleInstaller.cs
@@ -1,6 +1,6 @@
 // Copyright (c) .NET Foundation and contributors. All rights reserved. Licensed under the Microsoft Reciprocal License. See LICENSE.TXT file in the project root for full license information.
 
-namespace WixToolsetTest.BurnE2E
+namespace WixTestTools
 {
     using System;
     using System.IO;
