+++ src/WixToolsetTest.BurnE2E/FailureTests.cs
@@ -2,6 +2,7 @@
 
 namespace WixToolsetTest.BurnE2E
 {
+    using WixTestTools;
     using Xunit;
     using Xunit.Abstractions;
